backend/requirements.txt
fastapi
uvicorn
python-multipart
opencv-python
numpy
pillow
ultralytics
torch
torchvision
transformers
detectron2
pymupdf
pdf2image